Ref: Refactor checkpoint_validation.py --> validation_util.py - Now it has 2 different  modes VALIDATION_METHOD=["save","restore"]. - Add Flag DELETE_POD will trigger deletion of head pod for pathways. Nt ready for McJax yet. - Now interruption workflow is Training --> Interruption(Delete pod) --> Check if restore from correct checkpoint.
