+++ .github/workflows/pw-single-slice-training-restore.yaml
@@ -5,18 +5,19 @@ jobs:
   pw-tpu-single-slice-training-test-job:
     runs-on: jobset # Runs on CPU because this schedules a JobSet that's later routed to TPU
     container:
-      image: us-central2-docker.pkg.dev/tpu-prod-env-one-vm/camilo-tpu/pw-arc-jobset:latest
+      image: us-central2-docker.pkg.dev/tpu-prod-env-one-vm/camilo-tpu/pw-arc-jobset-val:latest
       env:
         # Required environmental variables for a successful run
         ARC_JOBSET_NAME: arc-pw-training
         ARC_JOBSET_JSON: /var/arc/arc-pw.json
         GCS_PREFIX: gs://axlearn-arc-testing/testing
-        JOBSET_DOCKER_IMAGE: us-central2-docker.pkg.dev/tpu-prod-env-one-vm/camilo-tpu/pw-arc-python:latest
+        JOBSET_DOCKER_IMAGE: us-central2-docker.pkg.dev/tpu-prod-env-one-vm/camilo-tpu/pw-arc-python-val:latest
         JOBSET_HEALTHY_TIMEOUT: 900
         ACCELERATOR: v6e_4x4_1
         # Optional flags
         STEPS_CHECKPOINT: 100
         MAX_STEPS: 500
+        RESTORE_MODE: True
         # CUSTOM_GIT_ORIGIN: https://github.com/CIeNET-International/axlearn # Optionally point to another repo
         # CUSTOM_GIT_BRANCH: cienet_jax_py3.12_nightly
         # Optional flags
@@ -47,8 +48,9 @@ jobs:
           GH_RUN_ID: ${{ github.run_id }}
           START_TIME: ${{ steps.launch_step.outputs.start_time}}
           END_TIME: ${{ steps.launch_step.outputs.end_time}}
+          VALIDATION_METHOD: save
         run: |
-          python /var/arc/checkpoint_validation.py
+          python /var/arc/validation_utils.py
       - name: Upload the training result to GCS
         if: always()
         run: GH_RUN_ID=${{ github.run_id }} bash /var/arc/upload-result.sh
@@ -56,17 +58,32 @@ jobs:
   interruption-jobset:
     runs-on: jobset # Runs on CPU because this schedules a JobSet that's later routed to TPU
     container:
-      image: us-central2-docker.pkg.dev/tpu-prod-env-one-vm/camilo-tpu/pw-arc-jobset:latest
+      image: us-central2-docker.pkg.dev/tpu-prod-env-one-vm/camilo-tpu/pw-arc-jobset-val:latest
       env:
         # Required environmental variables for a successful run
-        JOBSET_HEALTHY_TIMEOUT: 600
-        ARC_JOBSET_NAME_TO_CANCEL: arc-pw-training # Same as launch task.
+        JOBSET_HEALTHY_TIMEOUT: 1000
+        CUSTOM_GIT_BRANCH: main # Customize with a different branch
+        GCS_PREFIX: gs://axlearn-arc-testing/testing
     steps:
-      - name: Install GKE Auth Plugin
+      - name: Install the Kubernetes Python SDK
         run: |
-          apt-get update
-          apt-get install -y kubectl google-cloud-cli-gke-gcloud-auth-plugin
-          kubectl version --client
-      - name: Launch the Interruption job
+          uv pip install kubernetes
+      - name: Launch Pod deletion
+        id: launch_deletion
+        env:
+          GH_RUN_ID: ${{ github.run_id }}
+          ARC_JOBSET_NAME_TARGET: arc-pw-training
+          DELETE_POD: True
         run: |
-          GH_RUN_ID=${{ github.run_id }} bash /var/arc/arc-interruption-script.sh
+          START_TIME=$(date +%s)
+          echo "START_TIME=$START_TIME" >> $GITHUB_ENV
+          START_TIME=$START_TIME GH_RUN_ID=${{ github.run_id }} python /var/arc/validation_utils.py
+          echo "start_time=$START_TIME" >> $GITHUB_OUTPUT
+      - name: Launch the restore Validation
+        env:
+          GH_RUN_ID: ${{ github.run_id }}
+          START_TIME: ${{ steps.launch_deletion.outputs.start_time}}
+          ARC_JOBSET_NAME_TARGET: arc-pw-training
+          VALIDATION_METHOD: restore
+        run: |
+          START_TIME=$START_TIME GH_RUN_ID=${{ github.run_id }} python /var/arc/validation_utils.py
