+++ .github/workflows/pw-single-slice-training-test.yaml
@@ -5,13 +5,13 @@ jobs:
   pw-tpu-single-slice-training-test-job:
     runs-on: jobset # Runs on CPU because this schedules a JobSet that's later routed to TPU
     container:
-      image: us-central2-docker.pkg.dev/tpu-prod-env-one-vm/camilo-tpu/pw-arc-jobset:latest
+      image: us-central2-docker.pkg.dev/tpu-prod-env-one-vm/camilo-tpu/pw-arc-jobset-val:latest
       env:
         # Required environmental variables for a successful run
         ARC_JOBSET_NAME: arc-pw-training
         ARC_JOBSET_JSON: /var/arc/arc-pw.json
         GCS_PREFIX: gs://axlearn-arc-testing/testing
-        JOBSET_DOCKER_IMAGE: us-central2-docker.pkg.dev/tpu-prod-env-one-vm/camilo-tpu/pw-arc-python:latest
+        JOBSET_DOCKER_IMAGE: us-central2-docker.pkg.dev/tpu-prod-env-one-vm/camilo-tpu/pw-arc-python-val:latest
         JOBSET_HEALTHY_TIMEOUT: 900
         ACCELERATOR: v6e_4x4_1
         # Optional flags
@@ -48,26 +48,9 @@ jobs:
           GH_RUN_ID: ${{ github.run_id }}
           START_TIME: ${{ steps.launch_step.outputs.start_time}}
           END_TIME: ${{ steps.launch_step.outputs.end_time}}
+          VALIDATION_METHOD: save
         run: |
-          python /var/arc/checkpoint_validation.py
+          python /var/arc/validation_utils.py
       - name: Upload the training result to GCS
         if: always()
         run: GH_RUN_ID=${{ github.run_id }} bash /var/arc/upload-result.sh
-
-  interruption-jobset:
-    runs-on: jobset # Runs on CPU because this schedules a JobSet that's later routed to TPU
-    container:
-      image: us-central2-docker.pkg.dev/tpu-prod-env-one-vm/camilo-tpu/pw-arc-jobset:latest
-      env:
-        # Required environmental variables for a successful run
-        JOBSET_HEALTHY_TIMEOUT: 600
-        ARC_JOBSET_NAME_TO_CANCEL: arc-pw-training # Same as launch task.
-    steps:
-      - name: Install GKE Auth Plugin
-        run: |
-          apt-get update
-          apt-get install -y kubectl google-cloud-cli-gke-gcloud-auth-plugin
-          kubectl version --client
-      - name: Launch the Interruption job
-        run: |
-          GH_RUN_ID=${{ github.run_id }} bash /var/arc/arc-interruption-script.sh
